Move ShpD2 sprite loading into Mods.Common.
